requirements.txt
appdirs==1.4.4
black==20.8b1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
facenet-pytorch==2.5.2
flake8==3.9.1
idna==2.10
joblib==1.0.1
mccabe==0.6.1
mypy-extensions==0.4.3
numpy==1.20.2
opencv-python==4.5.1.48
pathspec==0.8.1
Pillow==8.2.0
pycodestyle==2.7.0
pyflakes==2.3.1
regex==2021.4.4
requests==2.25.1
scikit-learn==0.24.1
scipy==1.6.2
threadpoolctl==2.1.0
toml==0.10.2
torch==1.8.1
torchvision==0.9.1
typed-ast==1.4.3
typing-extensions==3.7.4.3
urllib3==1.26.4
